% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{AdjustDataToReflectWeights}
\alias{AdjustDataToReflectWeights}
\title{\code{AdjustDataToReflectWeights}}
\usage{
AdjustDataToReflectWeights(data, weights, seed = 123, silent = FALSE)
}
\arguments{
\item{data}{A \code{\link{data.frame}}.}

\item{weights}{The sampling or replication weights.}

\item{seed}{The seed used in random number generation. If \code{NULL}, the seed is not set.}

\item{silent}{If  \code{FALSE}, a warning is given if the weights are non-integers.}
}
\description{
Creates a new \code{\link{data.frame}} to reflect weights.
}
\details{
In situations where an algorithm does not accomodate weights, this
function modifies the \code{\link{data.frame}} by either: (A) stretching it
out, where the the weights are integers, or (B) resampling to create a new
bootstrapped \code{\link{data.frame}}, where the \code{weights} are
proportional to the probability of selection. When creating the bootstrap
sample, the sample size is whichever is greatest of the rounded sum and 1.
# Inspired by Zelig, 13-11-15.
}
