% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asdataframe.R
\name{AsDataFrame}
\alias{AsDataFrame}
\title{\code{AsDataFrame}}
\usage{
AsDataFrame(input.data, use.names = FALSE, ignore.columns = "",
  categorical.as.binary = FALSE, remove.first = FALSE)
}
\arguments{
\item{input.data}{Either a \code{\link{data.frame}}, a \code{\link{list}} of
\code{\link{data.frame}}s and/or \code{\link{vector}}s, or a \code{\link{matrix}}.}

\item{use.names}{Whether to use names in place of labels, used only for a
\code{\link{list}} of \code{\link{vector}}s.}

\item{ignore.columns}{A list of names of columns to ignore. When \code{input.data}
is a \code{\link{matrix}}, rows are also ignored.}

\item{categorical.as.binary}{Whether to convert factors to dummy binary variables,
or else their levels are converted to integers.}

\item{remove.first}{Whether to remove the first binary variable.}
}
\description{
Converts input data to a numeric \code{\link{data.frame}}.
}
