% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.R
\name{CreatingBinaryDependentVariableIfNecessary}
\alias{CreatingBinaryDependentVariableIfNecessary}
\title{\code{CreatingBinaryDependentVariableIfNecessary}}
\usage{
CreatingBinaryDependentVariableIfNecessary(formula, data)
}
\arguments{
\item{formula}{A formula.}

\item{data}{A data.frame}
}
\description{
Dichotomizes the dependent variable in a data.frame if not already dichotomized.
}
