% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.R
\name{CreatingFactorDependentVariableIfNecessary}
\alias{CreatingFactorDependentVariableIfNecessary}
\title{Turn the outcome variable from a formula into a factor}
\usage{
CreatingFactorDependentVariableIfNecessary(formula, data)
}
\arguments{
\item{formula}{A formula.}

\item{data}{A data.frame containing the variable.}
}
\description{
if not already a factor.
}
