% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.R
\name{FactorToIndicators}
\alias{FactorToIndicators}
\title{\code{FactorsToIndicators}}
\usage{
FactorToIndicators(variable, variable.name = deparse(substitute(variable)))
}
\arguments{
\item{variable}{The factor variable to convert.}

\item{variable.name}{The name of the input variable.}
}
\description{
Convert a factor variable to a matrix whose columns are binary variables
representing one of the levels from the factor variable.
}

