% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{ListToDataFrame}
\alias{ListToDataFrame}
\title{\code{ListToDataFrame}}
\usage{
ListToDataFrame(list.of.variables, binary = TRUE)
}
\arguments{
\item{list.of.variables}{A list containing the variables to combine. The
elements of the list should be of class numeric, factor, or ordered factor.}

\item{binary}{If \code{TRUE}, unordered factors are represented as dummy variables.
Otherwise, they are represented as sequential integers.}
}
\description{
Coerce a list of numeric or factor variables into a data frame.
}

