% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nicenumericcuts.R
\name{NiceNumericCuts}
\alias{NiceNumericCuts}
\title{Create factor variables from numeric variables using one of a variety of
methods.}
\usage{
NiceNumericCuts(
  input.data,
  method = c("tidy.intervals", "percentiles", "equal.width", "custom"),
  num.categories = 2,
  right = TRUE,
  round.input.data = FALSE,
  decimals = 1,
  label.decimals = 1,
  open.ends = TRUE,
  label.style = "tidy.labels",
  number.prefix = "",
  number.suffix = "",
  open.bottom.string = "Less than ",
  closed.bottom.string = " and below",
  open.top.string = "More than ",
  closed.top.string = " and over",
  equal.intervals.start = "",
  equal.intervals.end = "",
  equal.intervals.increment = "",
  custom.breaks = "",
  custom.always.includes.endpoints = FALSE,
  percents = "",
  quantile.type = 7,
  factors.use.labels = TRUE,
  grouping.mark = ",",
  decimals.mark = "."
)
}
\arguments{
\item{input.data}{The \code{\link{data.frame}} or \code{\link{vector}} 
containing the data to be categorized. Data should be numeric.}

\item{method}{A string describing which method is to be used to categorize
the data. Options are \code{tidy.intervals}, \code{percentiles},
\code{equal.width}, and \code{custom}.}

\item{num.categories}{An integer sepcifying the number of categories for
the new factor. This does not apply for \code{percentiles} or for
\code{custom} because in those cases the number of categories is
implied by the \code{percents} and \code{custom.breaks} arguments
respectively.}

\item{right}{A boolean value specifying that when determining breaks
in the data, the close side of the inerval should be on the right 
(i.e on the larger end of the interval).}

\item{round.input.data}{A boolean value specifying whether the input
data should be rounded before categorization.}

\item{decimals}{An integer which determines to how many decimals the}

\item{label.decimals}{An integer value which determines how many
decimal places should be included in formatted labels.}

\item{open.ends}{A boolean value which determines if labels at the upper
and lower ends of the range should be open-ended or should contain
both end points of the interval. E.g. if \code{TRUE} then interval
(0, 17] would be "Less than 18", otherwise it would be "0 to 17".}

\item{label.style}{A character indicating the style of labels to use
for the new factor levels. Options are \code{tidy.labels}, 
\code{inequality.notation}, \code{interval.notation}, and
\code{percentiles}, with the latter option only being applicable
if \code{method} is also \code{percentiles}.}

\item{number.prefix}{A character to be appended before numbers in new
factor labels.}

\item{number.suffix}{A character to be after before numbers in new
factor labels.}

\item{open.bottom.string}{A character indicating text to be placed at the
beginning of open-ended labels at the start of the range for open
intervals.}

\item{closed.bottom.string}{A character indicating text to be placed at the
end of open-ended labels at the start of the range for closed
intervals.}

\item{open.top.string}{A character indicating text to be placed at the
beginning of open-ended labels at the end of the range for open
intervals.}

\item{closed.top.string}{A character indicating text to be placed at the
end of open-ended labels at the end of the range for closed
intervals.}

\item{equal.intervals.start}{A numeric value indicating the start of the
range when using \code{method} of \code{equal.width}.}

\item{equal.intervals.end}{A numeric value indicating the end of the
range when using \code{method} of \code{equal.width}.}

\item{equal.intervals.increment}{A numeric value specifying the width of
the increments when using \code{method} of \code{equal.width}. This
value is optional, and overrides \code{num.categories} as the number
of categories is determined using the start, end, and increment.}

\item{custom.breaks}{A character containing a comma-seprated list of 
numeric values to be used as custom break points when the 
\code{method} is \code{custom}.}

\item{custom.always.includes.endpoints}{A logical value indicating whether,
when \code{method} is \code{custom}, the endpoints are always included.
If \code{FALSE}, the endpoints for the cuts are determined by the
\code{custom.breaks} only.}

\item{percents}{A single numeric value, or a character containing a comma
-separated list of numeric values to be used when the \code{method}
of \code{percentiles} is used. Values should be between 0 and 100.}

\item{quantile.type}{An interger between 1 and 9 to be passed to
\code{quantile} which determines the algorithm for creating quantiles.}

\item{factors.use.labels}{A logical value indicating whether numeric
information should be extracted from factor labels. If \code{FALSE}
the function will instead try to extract the underlying numerice
values from Q/Displayr.}

\item{grouping.mark}{A character to be used as the thousands-grouping
mark when inferring numeric information from factor labels.}

\item{decimals.mark}{A character to be used as the decimals-grouping
mark when inferring numeric information from factor labels.}
}
\value{
A data frame containing the new factor variables as columns.
}
\description{
Create factor variables from numeric variables using one of a variety of
methods.
}
