% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalevariableset.R
\name{numbersFromCategoricalVariableSets}
\alias{numbersFromCategoricalVariableSets}
\title{Get numbers from attributtes of Pick One and Pick One - Multi questions}
\usage{
numbersFromCategoricalVariableSets(x)
}
\arguments{
\item{x}{A variable set from Q/Displayr (an R factor or data.frame
with special attributes).}
}
\value{
For data.frame (vector) \code{x}, A matrix (vector) of
    numeric values with number of rows (elements) equal to
    \code{NROW(x)}.
}
\description{
Extracts the codeframe and value attributes from the (R
 data.frame/factor representation of) Displayr's Categorical
 Variable Sets (PickOne and PickOneMulti questions in Q) and
 applies them to the input factors/data.frame to convert it to its
 underlying values.
}
\details{
levels(x) are always present in the names of the codeframe
    attribute, \emph{unless} codes are hidden. The codeframe
    contains sourcevalues for each code/level, which need to be
    mapped from the sourcevalues attribute (original values as read
    in with the data set) to the (possibly user-modified) values
    attribute.

    A level that is the result of a merge will contain multiple
    source values for that element in the codeframe. These are
    mapped to the corresponding underlying values in the values
    attribute and averaged.
}
\keyword{internal}
