% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asnumeric.R
\name{OneHot}
\alias{OneHot}
\title{\code{OneHot}}
\usage{
OneHot(data, outcome = NULL)
}
\arguments{
\item{data}{A \code{\link{data.frame}}.}

\item{outcome}{Optionally, the name of a variable in \code{data} to be converted to a numeric vector.}
}
\description{
Produces a numeric \code{\link{matrix}} of binary predictor variables and a numeric \code{\link{vector}}
of an outcome variable from an input \code{\link{data.frame}}.
With the exception of \code{outcome}, all \code{\link{factor}} variables of \code{data} are
converted to one binary column per factor level.  Numeric variables are unchanged.
}
