% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.R
\name{Ordered}
\alias{Ordered}
\title{\code{Ordered}}
\usage{
Ordered(x, ...)
}
\arguments{
\item{x}{A vector of data, usually taking a small number of distinct values.}

\item{...}{Further arguments passed to \code{factor}.}
}
\description{
Converts to an ordered, but retains the label attribute.
}
\details{
If the variable is already a factor, nothing is done to it.
}

