% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataentry.R
\name{ParseAsDataFrame}
\alias{ParseAsDataFrame}
\title{\code{ParseEnteredData}}
\usage{
ParseAsDataFrame(m, warn = TRUE, want.factors = FALSE,
  want.col.names = TRUE, want.row.names = FALSE, us.format = TRUE)
}
\arguments{
\item{m}{Data frame which requires parsing.}

\item{warn}{Whether to show warnings.}

\item{want.factors}{Whether a text variable should be converted to a factor in a data frame.}

\item{want.col.names}{Whether to interpret the first row as column names in a data frame.}

\item{want.row.names}{Whether to interpret the first col as row names in a data frame.}

\item{us.format}{Whether to use the US convention when parsing dates in a data frame.}
}
\description{
Takes a data.frame and performs extra parsing, for example with dates and percentages.
}
