% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime.R
\name{ParseDateTime}
\alias{ParseDateTime}
\title{\code{ParseEnteredData}}
\usage{
ParseDateTime(x, us.format = TRUE)
}
\arguments{
\item{x}{Character vector to be parsed.}

\item{us.format}{Whether to use the US convention for dates.}
}
\description{
\code{ParseEnteredData}
}

