% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataentry.R
\name{ParseEnteredData}
\alias{ParseEnteredData}
\title{\code{ParseEnteredData}}
\usage{
ParseEnteredData(raw.matrix)
}
\arguments{
\item{raw.matrix}{Character matrix}
}
\description{
Takes a raw character matrix returned by dataEntry R GUI control
and attempts to parse it to something more friendly such as a numeric matrix.
}

