% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsetext.R
\name{ParseText}
\alias{ParseText}
\title{Parse text into the appropriate type}
\usage{
ParseText(x, same.as = NULL, type = "Automatic")
}
\arguments{
\item{x}{The input character string to be converted.}

\item{same.as}{A variable of the same class as the desired output. If this is supplied then \code{class(same.as)} will be used instead of \code{type}.}

\item{type}{The desired type of the output. If set to "Automatic" 
(and \code{same.as} is not supplied) then \code{numeric}, \code{POSIXct}, 
\code{Date}, and \code{character} will be tried sequentially. However,
if a specific type is specified, then \code{NA} will be returned if
it cannot be appropriately parsed. Note that \code{factors} will be 
treated as characters}
}
\description{
Convert text into the specified type. This can be used with \code{form.textBox}.
The return value is a single value. Use \link{TextAsVector} first if multiple
entries separated by a comma (or other deliminator) is expected.
}
