% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseUserEnteredTable.R
\name{ParseUserEnteredTable}
\alias{ParseUserEnteredTable}
\title{Parse data from dataEntry R GUI control}
\usage{
ParseUserEnteredTable(raw.matrix, warn = TRUE, want.data.frame = FALSE,
  want.factors = TRUE, want.col.names = TRUE, want.row.names = FALSE,
  us.format = NULL)
}
\arguments{
\item{raw.matrix}{Character matrix}

\item{warn}{Whether to show warnings}

\item{want.data.frame}{logical; should a \code{data.frame} be returned instead
of a matrix or vector? If the input matrix returns a text matrix, this converted
to a dataframe}

\item{want.factors}{logical; should a text variable be converted to a factor?
Ignored if \code{want.data.frame} is \code{FALSE}}

\item{want.col.names}{logical; should the first row be interpretted as column names?
Ignored if \code{want.data.frame} is \code{FALSE}}

\item{want.row.names}{logical; should the first colulm be interpretted as row names?
Ignored if \code{want.data.frame} is \code{FALSE}}

\item{us.format}{logical; should the U.S. convention be used when parsing dates?
Ignored if \code{want.data.frame} is \code{FALSE}}
}
\value{
if \code{want.data.frame == FALSE}, numeric vector or matrix, possibly
with an attribute \code{"statistic"} if row and column names are present; otherwise,
a \code{data.frame}
}
\description{
Takes a raw character matrix returned by dataEntry R GUI control
and attempts to parse it to something more friendly such as a
numeric matrix.
}
\details{
First removes empty rows and columns.  If every entry of
    the resulting matrix (or vector) is numeric, a matrix without
    row or column names is returned.  If the matrix has row and
    column names and a value in the \code{[1,1]} entry, that entry
    will be returned as an attribute in the output called
    \code{"statistic"}.
}
\note{
If characters are present in the entire first row (including
    the \code{[1,1]} entry), but the first column has all numeric
    entries aside from that \code{[1,1]} entry, then the resulting matrix
will have column names, but no row names.  And similarly for the case when
the first row is all numeric, but the \code{[1,1]} entry is not, the output
matrix will have row names, but no column names.
    to extract both row and column names from the resulting matrix
}
