% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackingutilities.R
\name{ProcessAndStackDataForRegression}
\alias{ProcessAndStackDataForRegression}
\title{Helper function to stack data for a regression model.}
\usage{
ProcessAndStackDataForRegression(
  unstacked.data,
  formula,
  interaction,
  subset,
  weights
)
}
\arguments{
\item{unstacked.data}{A named list of two data frames. Element "X" is the data frame containing
containing the outcome variable (usually a Numeric - Multi or Nominal/Ordinal - Multi 
variable set in Displayr), and element "Y" is the data frame containing the predictor
variables (a Binary - Grid or Numeric - Grid variable set in Displayr).}

\item{formula}{A formula object for the regression.}

\item{interaction}{Optional variable to test for interaction with other variables in the model.}

\item{subset}{An optional vector specifying a subset of observations to be
used in the fitting process.}

\item{weights}{An optional vector of sampling weights.}
}
\description{
Takes the input unstacked data, interaction, subset, weights and formula terms,
processes the unstacked data, and stacks it.
If the stacking is successful, the interaction, subset, weights are also updated
to be the appropriate size
}
