% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.R
\name{ProcessQVariables}
\alias{ProcessQVariables}
\title{ProcessQVariables}
\usage{
ProcessQVariables(x)
}
\arguments{
\item{x}{A Q variable or a data frame containing Q variables.}
}
\description{
Processes Q variables, e.g.: converting date variables to be categorical
based on their period. This function should be called to process Q variables before they are used.
}
