% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{QuestionListToDataFrame}
\alias{QuestionListToDataFrame}
\title{\code{QuestionListToDataFrame}}
\usage{
QuestionListToDataFrame(list.of.questions, names.to.remove = c("NET", "Total",
  "SUM"))
}
\arguments{
\item{list.of.questions}{A list containing the questions to combine.}

\item{names.to.remove}{A vector of column names to exclude.}
}
\description{
Combine a list of questions into a data frame.
}
