% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removebynames.R
\name{RemoveByName}
\alias{RemoveByName}
\title{Remove indices from a vector by name}
\usage{
RemoveByName(x, rnames)
}
\arguments{
\item{x}{named vector, data.frame, or list to remove entries from}

\item{rnames}{either 1) a character, vector giving entry names to remove; 2) a single
string containing comma or semi-colon separated names to remove, 3) a list where each
element is either 1) or 2)}
}
\value{
\code{x} with entries specified in \code{rnames} removed
they will be combined using \code{\link{union}}
}
\description{
Remove elements from a vector while preserving attributes
}
\details{
Will also work for data.frame \code{x}
}
\note{
An error is thrown if removal would result in an empty vector
}
\examples{
x <- c(a = 1, b = 2, c = 3, d = 4, e = 5)
RemoveByName(x, "a")
RemoveByName(x, "a; b")
RemoveByName(x, list(c("a", "b"), " c,  d"))
}
