% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{RemoveRowsAndOrColumns}
\alias{RemoveRowsAndOrColumns}
\title{\code{RemoveRowsAndOrColumns}}
\usage{
RemoveRowsAndOrColumns(x, row.names.to.remove = c("NET", "Total", "SUM"),
  column.names.to.remove = c("NET", "Total", "SUM"))
}
\arguments{
\item{x}{The data that is being analyzed}

\item{row.names.to.remove}{A vector of the row labels to remove.}

\item{column.names.to.remove}{A vector of the column labels to remove.}
}
\description{
Removes rows or columns from the table.
}

