% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{RetainedRowsAndOrColumns}
\alias{RetainedRowsAndOrColumns}
\title{\code{RetainedRowsAndOrColumns}}
\usage{
RetainedRowsAndOrColumns(x, row.names.to.remove = c("NET", "Total", "SUM"),
  column.names.to.remove = c("NET", "Total", "SUM"), split = ",")
}
\arguments{
\item{x}{The data that is being analyzed.}

\item{row.names.to.remove}{A vector or comma-separated string containing the row labels to remove.}

\item{column.names.to.remove}{A vector or comma-separated string containing the column labels to remove.}

\item{split}{Delimiter to split string on.}
}
\description{
Produces a list of 2 vectors, indices of rows and columns from a table.
}
\details{
Trailing spaces are removed and lower/upper case is ignored.
}
