% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalevariableset.R
\name{ScaleVariableSet}
\alias{ScaleVariableSet}
\title{Scale a Displayr Variable Set}
\usage{
ScaleVariableSet(
  data,
  type = c("standardize", "center", "unit"),
  within.case = FALSE
)
}
\arguments{
\item{data}{A Displayr Variable Set or Q Question: A numeric or
factor vector, or a data.frame containing either all numeric
columns or all factor columns.}

\item{type}{Character string specifying the type of scaling to
perform; one of \code{"standardize"} (the default),
\code{"center"}, or \code{"unit"}.}

\item{within.case}{Logical; should the scaling be performed with
case (rows) or variables (columns). An error is through if
\code{TRUE} and \code{data} contains only one variable.}
}
\value{
A numeric vector if \code{data} contains only a single
    variable or a data.frame containing the scaled values.
}
\description{
Given a Displayr Variable Set or Q Question, performs a specified
scaling on the data; first converting any categorical variables to
numeric as necessary.
}
\examples{
x <- structure(c(1, 2, 3, 4, 5), questiontype = "Number",
               dataset = "dat", name = "x", label = "x", question = "x")

ScaleVariableSet(x, type = "unit")
}
\seealso{
\code{\link{scale}}
}
