% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{SplitVectorToList}
\alias{SplitVectorToList}
\title{\code{SplitVectorToList}}
\usage{
SplitVectorToList(values, groups)
}
\arguments{
\item{values}{A \code{\link{vector}} to be split.}

\item{groups}{A \code{\link{vector}} which determines how \code{vakuees} is split.}
}
\value{
A \code{\link{list}}.
}
\description{
Splits a vector \code{values} by unique values of another vector, \code{groups}.
}
