% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackingutilities.R
\name{StackTextAndCategorization}
\alias{StackTextAndCategorization}
\title{Stacks several text variables and an existing categorization 
in a consisten manner for text analysis functions.}
\usage{
StackTextAndCategorization(
  text,
  existing.categorization = NULL,
  subset = TRUE,
  weights = NULL
)
}
\description{
Stacks several text variables and an existing categorization 
in a consisten manner for text analysis functions.
}
