% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{StandardizeData}
\alias{StandardizeData}
\title{StandardizeData}
\usage{
StandardizeData(data, method)
}
\arguments{
\item{data}{A \code{data.frame} or \code{matrix}.}

\item{method}{The standardization method. Takes values \code{"z-scores"}, \code{"Range [-1,1]"},
\code{"Range [0,1]"}, \code{"Maximum magnitude of 1"}, \code{"Mean of 1"} and \code{"Standard deviation of 1"}.}
}
\description{
StandardizeData
}

