% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataentry.R
\name{TextAsVector}
\alias{TextAsVector}
\title{\code{TextAsVector}}
\usage{
TextAsVector(x, split = ",", silent = FALSE)
}
\arguments{
\item{x}{Input text, which may be either a deliminated string which is broken up
or a vector of strings which need to be cleaned up.}

\item{split}{Deliminator to split input text.}

\item{silent}{Boolean indicating whether a warning is given if smart quotes are removed}
}
\description{
Cleans up input text into a vector of strings. The input text is split
   using the specified deliminater, smart quotes removed and trailing and leading
   white space and quotes removed.
}
