% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.R
\name{Unclass}
\alias{Unclass}
\title{Unclasses and removes the levels of a factor.}
\usage{
Unclass(x, warn = TRUE, use.numeric.levels = TRUE)
}
\arguments{
\item{x}{A factor.}

\item{warn}{Show warning.}

\item{use.numeric.levels}{If \code{TRUE}, numeric factor levels are used as the values of the output,
else levels are output as 1, 2, 3, ...}
}
\description{
Unclasses and removes the levels of a factor.
}
