% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.R
\name{UnclassIfNecessary}
\alias{UnclassIfNecessary}
\title{\code{UnclassIfNecessary}}
\usage{
UnclassIfNecessary(x, warn = TRUE)
}
\arguments{
\item{x}{A vector.}

\item{warn}{Show warning.}
}
\value{
A vector
}
\description{
Unclasses a variable if it is a factor. Otherwise, returns x.
}
