% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackingutilities.R
\name{UnstackCategorization}
\alias{UnstackCategorization}
\title{Unstack a data frame which corresponds to a categorization of mulitple
text variables in Displayr.}
\usage{
UnstackCategorization(categorization, inds)
}
\arguments{
\item{categorization}{The data frame to be unstacked. Its columns are factors
for a single-response categorization (Pick One - Multi) or binary numeric
vectors for a multiple-response categorization (Pick Any - Grid)}

\item{inds}{A character vector whose entries are of the form "<casenumber>.<variablename>"
such as would be produced by the \code{stack} function. Usually would be
generated by when originally stacking the text data via \code{StackTextAndCategorization}}
}
\description{
Unstack a data frame which corresponds to a categorization of mulitple
text variables in Displayr.
}
