% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svd.R
\name{WeightedSVD}
\alias{WeightedSVD}
\title{Weighted Singular Value Decomposition
\code{WeightedSVD} Computes a SVD with frequency weights.}
\usage{
WeightedSVD(x, weights = rep(1, n), nu = min(n, p), nv = min(n, p))
}
\arguments{
\item{x}{A numeric or complex matrix whose SVD decomposition is to be computed. Logical matrices are coerced to numeric.}

\item{weights}{Frequency Weights.}

\item{nu}{The number of left singular vectors to be computed. This must between 0 and n = nrow(x).}

\item{nv}{The number of right singular vectors to be computed. This must between 0 and n = ncol(x).}
}
\description{
Weighted Singular Value Decomposition
\code{WeightedSVD} Computes a SVD with frequency weights.
}
