% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asnumeric.R
\name{AsNumeric}
\alias{AsNumeric}
\title{\code{AsNumeric}}
\usage{
AsNumeric(x, binary = TRUE, name = NULL, remove.first = FALSE)
}
\arguments{
\item{x}{The \code{\link{data.frame}} or \code{\link{vector}}.}

\item{binary}{If \code{TRUE}, unordered factors are represented as dummy variables.
Otherwise, they are represented as sequential integers.}

\item{name}{Used if \code{binary} is \code{TRUE} to construct variable names. This parameter is
ignored if x is a \code{\link{data.frame}} or a \code{\link{list}}.}

\item{remove.first}{Remove the first binary variable.}
}
\description{
Coerces a \code{\link{data.frame}} or \code{\link{vector}} that may contain
factors into a \code{\link{vector}} or \code{\link{data.frame}} that does not contain factors.
}
\details{
Characters are treated as factors.
}
