% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.R
\name{CreatingBinaryVariableIfNecessary}
\alias{CreatingBinaryVariableIfNecessary}
\title{\code{CreatingBinaryVariableIfNecessary}}
\usage{
CreatingBinaryVariableIfNecessary(data, variable.name)
}
\arguments{
\item{data}{A data.frame}

\item{variable.name}{The name of the variable in the data.frame.}
}
\description{
Dichotomizes a variable.
}

