% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.R
\name{CreatingFactorDependentVariableIfNecessary}
\alias{CreatingFactorDependentVariableIfNecessary}
\title{\code{CreatingFactorDependentVariableIfNecessary}}
\usage{
CreatingFactorDependentVariableIfNecessary(formula, data)
}
\arguments{
\item{formula}{A formula.}

\item{data}{A data.frame}
}
\description{
Turns a numeric outcome variable into a factor outcome variable in a data.frame if not already a factor.
}

