% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.R
\name{FactorToNumeric}
\alias{FactorToNumeric}
\title{\code{FactorToNumeric}}
\usage{
FactorToNumeric(x, variable.name = deparse(substitute(x)))
}
\arguments{
\item{x}{A factor or ordered factor.}

\item{variable.name}{The name of the variable.}
}
\description{
Convert a factor variable to a numeric vector (when the factor is ordered),
or a matrix of indicator variables (when the factor is not ordered).
}

