% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.R
\name{FactorToNumeric}
\alias{FactorToNumeric}
\title{\code{FactorToNumeric}}
\usage{
FactorToNumeric(x, binary = TRUE,
  name = RemoveParentName(deparse(substitute(x))), remove.first = TRUE)
}
\arguments{
\item{x}{A factor or ordered factor.}

\item{binary}{Returns the factor as binary variables.}

\item{name}{The name of the variable.}

\item{remove.first}{Remove the first binary variable, if a binary variable is being created.}
}
\description{
Convert a factor variable to a numeric vector (when the factor is ordered),
or a matrix of indicator variables (when the factor is not ordered).
}

