% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergecategoriescontainingranges.R
\name{MergeRangeCategories}
\alias{MergeRangeCategories}
\title{Create factor new factor variables from existing factors whose
labels contain numeric ranges, like 18 to 24 or $2,000 to $3,000,}
\usage{
MergeRangeCategories(
  input.data,
  num.categories = 2,
  method = c("even.proportions", "even.ranges"),
  grouping.mark = ",",
  decimals.mark = "\\\\.",
  lower.bound = NULL,
  upper.bound = NULL
)
}
\arguments{
\item{input.data}{A \code{data.frame} or \code{vector} containing factors
to be merged.}

\item{num.categories}{An integer indicating the number of categories
in the new factors.}

\item{method}{A character indicating how the merging is to be performed.
The available options are \code{"even.proportions"} and 
\code{even.ranges}.}

\item{grouping.mark}{A character indicating how thousands are grouped within
the labels of the input data. Usally a comma, but sometimes a period
for example in European locales.}

\item{decimals.mark}{A character indicating how decimal values are separated
in the input data. Usually a period.}

\item{lower.bound}{A numeric value which indicates the lower bound of the
numeric ranges when the number is not explicitly included in the 
level labels. For example, when the lowest level is "Under 18", you
may wish to specify that the lowest possible value in this level
is 12, 15, 0, etc. This helps when you wish to}

\item{upper.bound}{As above, for the higher end of the range.}
}
\description{
Create factor new factor variables from existing factors whose
labels contain numeric ranges, like 18 to 24 or $2,000 to $3,000,
}
