% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.R
\name{OrderedToNumeric}
\alias{OrderedToNumeric}
\title{Convert an ordered factor to a numeric vector.}
\usage{
OrderedToNumeric(x)
}
\arguments{
\item{x}{An ordered factor.}
}
\description{
Convert an ordered factor to a numeric vector.
}
\details{
If all levels are numeric they are converted to the same numbers.
  Else they are converted to 1, 2, 3, . as per the ordering of the factors.
}
