% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataentry.R
\name{ParseAsDataFrame}
\alias{ParseAsDataFrame}
\title{Parse a Character Matrix To a Data Frame}
\usage{
ParseAsDataFrame(m, warn = TRUE, want.factors = FALSE,
  want.col.names = NULL, want.row.names = FALSE, us.format = NULL)
}
\arguments{
\item{m}{A character matrix to be converted to \code{data.frame}}

\item{warn}{Whether to show warnings.}

\item{want.factors}{Whether a text variable should be converted to a factor in a data frame.}

\item{want.col.names}{Whether to interpret the first row as column names in a data frame.}

\item{want.row.names}{Whether to interpret the first col as row names in a data frame.}

\item{us.format}{Whether to use the US convention when parsing dates in a data frame.}
}
\description{
Takes a data.frame and performs extra parsing, for example with
dates and percentages.
}
