% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataentry.R
\name{ParseEnteredData}
\alias{ParseEnteredData}
\title{\code{ParseEnteredData}}
\usage{
ParseEnteredData(raw.matrix, warn = TRUE, want.data.frame = FALSE,
  want.factors = TRUE, want.col.names = TRUE, want.row.names = FALSE,
  us.format = NULL)
}
\arguments{
\item{raw.matrix}{Character matrix}

\item{warn}{Whether to show warnings}

\item{want.data.frame}{logical; should a \code{data.frame} be returned instead
of a matrix or vector?}

\item{want.factors}{logical; should a text variable be converted to a factor?
Ignored if \code{want.data.frame} is \code{FALSE}}

\item{want.col.names}{logical; should the first row be interpretted as column names?
Ignored if \code{want.data.frame} is \code{FALSE}}

\item{want.row.names}{logical; should the first colulm be interpretted as row names?
Ignored if \code{want.data.frame} is \code{FALSE}}

\item{us.format}{logical; should the U.S. convention be used when parsing dates?
Ignored if \code{want.data.frame} is \code{FALSE}}
}
\description{
Takes a raw character matrix returned by dataEntry R GUI control
and attempts to parse it to something more friendly such as a numeric matrix.
}
