% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackingutilities.R
\name{StackTextAndCategorization}
\alias{StackTextAndCategorization}
\title{Stacks several text variables and an existing categorization 
in a consisten manner for text analysis functions.}
\usage{
StackTextAndCategorization(
  text,
  existing.categorization = NULL,
  subset = TRUE,
  weights = NULL
)
}
\arguments{
\item{text}{A data frame containing one or more character vectors. Typically
these start their life as variables in a Text - Multi variable set in Displayr
or are several related Text variable sets.}

\item{existing.categorization}{A data frame containing the corresponding
categories. This could be a data frame of factor variables (a single-response 
categorization), or a data frame of binary numeric variables (a multiple-response
categorization). This can be NULL if no existing categorization is to be used
in the analysis.}

\item{subset}{A boolean vector indicating the subset of cases that the user
wishes to include later in a classification of the text. Usually comes from QFilter.
This will be expanded to align with the stacked data.}

\item{weights}{A numeric vector of case weights for later us in the analysis.
This will be expanded to align with the stacked data.}
}
\description{
Stacks several text variables and an existing categorization 
in a consisten manner for text analysis functions.
}
