% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.R
\name{Unclass}
\alias{Unclass}
\title{\code{Unclass}}
\usage{
Unclass(x, warn = TRUE)
}
\arguments{
\item{x}{An ordered factor.}

\item{warn}{Show warning.}
}
\description{
Unclasses, and removes the levels attribute.
}
